requirements.txt

flask
playwright
google-api-python-client
google-auth

playwright==1.50.0
